requirements.txt
Flask==2.3.3
google-cloud-vision==3.7.2
